Oswaldo

+++ clase_1/code3_2_2.php
@@ -1,5 +1,5 @@
 <?php
-$a=(int) ( (0.1+0.7) * 10 );
+$a=( (0.1+0.7) * 10 );
 echo $a; // ¡muestra 7!
 echo "<br>";
 var_dump($a);
